+++ app/manager/createschedule/tempo_scheduling_webapp1.1_11.19.2024_working_calendar.code-workspace
@@ -1,0 +1,10 @@
+
+
+{
+	"folders": [
+		{
+			"path": "../../.."
+		}
+	],
+	"settings": {}
+}
